requirements.txt
Django==1.11.15
djangorestframework==3.8.2
psycopg2-binary==2.7.5
pytz==2018.5
pydub
google-cloud-storage
google-cloud-speech
